Make origin panel button one component for content
